requirements.txt
cachetools==5.3.1
deprecation==2.1.0
grpcio==1.56.0
packaging==23.1
protobuf==4.23.4
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b58
typing-extensions==4.7.1
ujson==5.8.0
